// src/interfaces/dataInterface.ts
// Generated by https://quicktype.io

export interface DataInterface {
  ok:      boolean;
  usuario: Usuario;
  token:   string;
  msg?:string
  mensajes?: Mensaje[]
}

export interface Usuario {
  nombre: string;
  email:  string;
  online: boolean;
  uid:     string;
}

export interface ChatState{
 
  chatActivo:string | null;
  usuarios:Usuario[];
  mensajes:Mensaje[];
}
export interface Mensaje{
  de:string
  para:string
  mensaje:string
  _id?:string
  createdAt:Date
  updatedAt:Date
}

// createdAt: "2023-04-01T21:11:26.844Z"
// de: "64288f150fba2d4dd03bc639"
// mensaje: "hola"
// para: "64288f010fba2d4dd03bc636"
// updatedAt: "2023-04-01T21:11:26.844Z"
// _id: "64289dfe474bf8f60aaf7c8b"